refactor(auth): extract proxy_auth seam
Move every proxy-auth primitive to a dedicated module:

- _PROXY_DEAD_PATHS (/setup, /login redirect targets in proxy mode)
- _is_proxy_auth_mode, _is_trusted_proxy, _extract_proxy_username
- _validate_proxy_auth (trust gate + header read in one call)
- _validate_proxy_csrf, _ensure_proxy_csrf_cookie (double-submit
  CSRF pattern scoped to proxy mode)

auth/__init__.py re-exports every name; the middleware that lives
there continues to call the helpers via bare name lookup (which
resolves to the re-exports), so monkeypatching
``houndarr.auth._is_trusted_proxy`` in tests still affects the
middleware's behaviour.  Phase 2h moves the middleware into its
own module and will update the patch targets accordingly.
